Initial commit for Replacing javascript.Net with V8.Net
This is a try to make the typecript addin faster and also
crossplattform win,mac,linux in the feature.
The speed up the analysis and compilation the scripts can be
precompiled, a possibility is to place them in a static class.
But there is maybe a better solution for this. With this commit we just
want to check if it is possible to replace Javascript.Net with V8.Net.