[FIX] [WIP] Update script content on text edit #5 #8
Currently the content of a script is not updated wenn editing this lead
to the situation that the native typescript service and monodevelop are
out of sync, monodevelop then search for dom regions that are not
available. Result index out of range exception.

Note:
Observer behavior of the parser is that on every character change it
tries to add the current file (AddFile()) to the host environment also it then tries
to add all files in the project to the host env. (AddFiles).
To Gain performance we should move the file add proccess elsewhere.