[WIP] Add implement change range function
Get change range is now recomputed every time when it gets called.

lib.d.ts is now not enymore part of the host scripts.
It is assumed that lib.d.ts does not get used by the langueage service
but is used by the compiler. The compiler can use the host function
"getDefaultLibScript".
Removing lib.d.ts has an improvement on the performance when calculating
getchangerange.

Note:
Implement a functionality that converts all files to line ending "\r\n" or "\n"
Typescripts ask when initializing for the type of lineendings, means we
can not use diffrent line endings in files.